+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added manifest hook for when CSS code splitting is off
+
 ## v0.1.2 (2022-12-12)
 * Added publicDir config
 * Added isHot flag to manifest